Crate async_modbus

async-modbus

async-modbus provides a lightweight implementation of Modbus requests and responses with the help of zerocopy. It is designed for resource-constrained environments (being no_std by default) like embedded systems but can be used in any Rust project.

There is a basic client implementation available behind the embedded-io feature flag (enabled by default).

Re-exports

pub use pdu::Pdu;

pub use zerocopy;

Modules

client

Client functions for embedded_io_async-based IO.

pdu

PDUs in Modbus are everything in between the unit ID and CRC.

Structs

Frame

A complete Modbus frame.

FrameBuilder

A builder for Frames.

FrameView

A frame with an unknown PDU type.

PduView

The most generic Modbus PDU, containing a function code and data payload.