async_modbus

Module pdu

PDUs in Modbus are everything in between the unit ID and CRC.

Modules

request

Modbus request PDUs. Wrap them in a Frame to give them a unit ID and CRC.

response

Response PDUs!

Structs

CrcError

Error indicating a CRC validation failure.

Enums

ValidationError

Errors that can occur when validating a Modbus response.

Traits

Pdu

A Protocol Data Unit (PDU) contains the function code and data.

Response

The Response trait is used to validate a response against a request, essentially to ensure that the response corresponds to the request.