async_modbus::pdu

Trait Response

pub trait Response<Request>: Pdu {
    type Data;

    // Required methods
    fn matches_request(&self, request: &Request) -> bool;
    fn into_data(self) -> Self::Data;
}

The Response trait is used to validate a response against a request, essentially to ensure that the response corresponds to the request.

It also provides a method to convert the response into its data payload.

Required Associated Types

type Data

The data payload of the response.

Required Methods

fn matches_request(&self, request: &Request) -> bool

Validates the response against the request.

fn into_data(self) -> Self::Data

Converts the response into its data payload.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.