async_modbus

Struct FrameView

pub struct FrameView<'a> { /* private fields */ }

A frame with an unknown PDU type.

let frame = FrameView::try_from_bytes(&hex!("01 06 00 04 00 02 49 CA")).unwrap();

assert_eq!(frame.unit_id(), 1);
assert!(frame.validate_crc().is_ok());

Implementations

impl<'a> FrameView<'a>

pub fn try_from_bytes(buf: &'a [u8]) -> Option<Self>

Parse a frame from a byte slice. This method does not validate the CRC or the PDU contents, only that the frame has a valid length.

pub const fn unit_id(&self) -> u8

The unit ID of the frame.

pub const fn crc(&self) -> u16

The CRC sent with the frame. This struct does not guarantee that the CRC is valid; be sure to check it with FrameView::validate_crc.

pub fn pdu(self) -> Result<&'a PduView, CrcError>

Returns the wrapped PDU if the CRC is valid.

pub fn validate_crc(&self) -> Result<(), CrcError>

Validate the CRC of the frame.

Note that FrameView::pdu validates the CRC before returning the PDU, rendering this method unnecessary to call directly in most cases.